+++ Microsoft.Azure.Cosmos/src/Linq/ICosmosLinqSerializer.cs
@@ -10,10 +10,19 @@ namespace Microsoft.Azure.Cosmos.Linq
 
     internal interface ICosmosLinqSerializer
     {
+        /// <summary>
+        /// Applies specified custom converters to an expression.
+        /// </summary>
         SqlScalarExpression ApplyCustomConverters(Expression left, SqlLiteralScalarExpression right);
 
-        string GetMemberName(MemberInfo memberInfo, CosmosLinqSerializerOptions linqSerializerOptions = null);
+        /// <summary>
+        /// Serializes a ConstantExpression as a SqlScalarExpression.
+        /// </summary>
+        SqlScalarExpression ConvertToSqlScalarExpression(ConstantExpression inputExpression, IDictionary<object, string> parameters);
 
-        SqlScalarExpression VisitConstant(ConstantExpression inputExpression, IDictionary<object, string> parameters);
+        /// <summary>
+        /// Gets a member name with any LINQ serializer options applied.
+        /// </summary>
+        string GetMemberName(MemberInfo memberInfo, CosmosLinqSerializerOptions linqSerializerOptions = null);
     }
 }
